controlador Juego test

+++ src/main/java/com/tallerwebi/dominio/RepositorioCuestionario.java
@@ -3,7 +3,7 @@ package com.tallerwebi.dominio;
 public interface RepositorioCuestionario {
 
     void guardar(Cuestionario cuestionario);
-    Cuestionario buscar(Integer id);
+    Cuestionario buscar(Long id);
     void modificar(Cuestionario cuestionario);
     Cuestionario buscarPorCategoria(String categoria);
     Cuestionario buscarPorDificultad(String dificultad);
